Factor out helpers from CBoostedTreeImpl
