+++ include/maths/CBoostedTreeImpl.h
@@ -7,28 +7,24 @@
 #ifndef INCLUDED_ml_maths_CBoostedTreeImpl_h
 #define INCLUDED_ml_maths_CBoostedTreeImpl_h
 
-#include <core/CContainerPrinter.h>
 #include <core/CDataFrame.h>
 #include <core/CImmutableRadixSet.h>
-#include <core/CLogger.h>
 #include <core/CMemory.h>
 #include <core/CPackedBitVector.h>
-#include <core/CSmallVector.h>
 #include <core/CStatePersistInserter.h>
 #include <core/CStateRestoreTraverser.h>
 
 #include <maths/CBasicStatistics.h>
 #include <maths/CBoostedTree.h>
 #include <maths/CBoostedTreeHyperparameters.h>
+#include <maths/CBoostedTreeUtils.h>
 #include <maths/CDataFrameAnalysisInstrumentationInterface.h>
 #include <maths/CDataFrameCategoryEncoder.h>
 #include <maths/CDataFrameUtils.h>
 #include <maths/CLinearAlgebraEigen.h>
 #include <maths/CPRNG.h>
-#include <maths/CTools.h>
 #include <maths/ImportExport.h>
 
-#include <boost/operators.hpp>
 #include <boost/optional.hpp>
 #include <boost/range/irange.hpp>
 
@@ -161,265 +157,17 @@ public:
     //@}
 
 private:
-    using TSizeDoublePr = std::pair<std::size_t, double>;
     using TDoubleDoublePr = std::pair<double, double>;
     using TOptionalDoubleVec = std::vector<TOptionalDouble>;
     using TOptionalDoubleVecVec = std::vector<TOptionalDoubleVec>;
     using TOptionalSize = boost::optional<std::size_t>;
-    using TImmutableRadixSetVec = std::vector<core::CImmutableRadixSet<double>>;
     using TVector = CDenseVector<double>;
-    using TRowItr = core::CDataFrame::TRowItr;
     using TPackedBitVectorVec = std::vector<core::CPackedBitVector>;
+    using TImmutableRadixSetVec = std::vector<core::CImmutableRadixSet<double>>;
     using TNodeVecVecDoublePr = std::pair<TNodeVecVec, double>;
     using TDataFrameCategoryEncoderUPtr = std::unique_ptr<CDataFrameCategoryEncoder>;
     using TDataTypeVec = CDataFrameUtils::TDataTypeVec;
     using TRegularizationOverride = CBoostedTreeRegularization<TOptionalDouble>;
-
-    //! \brief Maintains a collection of statistics about a leaf of the regression
-    //! tree as it is built.
-    //!
-    //! DESCRIPTION:\N
-    //! The regression tree is grown top down by greedily selecting the split with
-    //! the maximum gain (in the loss). This finds and scores the maximum gain split
-    //! of a single leaf of the tree.
-    class CLeafNodeStatistics final {
-    public:
-        CLeafNodeStatistics(std::size_t id,
-                            std::size_t numberInputColumns,
-                            std::size_t numberLossParameters,
-                            std::size_t numberThreads,
-                            const core::CDataFrame& frame,
-                            const CDataFrameCategoryEncoder& encoder,
-                            const TRegularization& regularization,
-                            const TImmutableRadixSetVec& candidateSplits,
-                            const TSizeVec& featureBag,
-                            std::size_t depth,
-                            const core::CPackedBitVector& rowMask);
-
-        //! Only called by split but is public so it's accessible to std::make_shared.
-        CLeafNodeStatistics(std::size_t id,
-                            std::size_t numberInputColumns,
-                            std::size_t numberLossParameters,
-                            std::size_t numberThreads,
-                            const core::CDataFrame& frame,
-                            const CDataFrameCategoryEncoder& encoder,
-                            const TRegularization& regularization,
-                            const TImmutableRadixSetVec& candidateSplits,
-                            const TSizeVec& featureBag,
-                            bool isLeftChild,
-                            std::size_t depth,
-                            const CBoostedTreeNode& split,
-                            const core::CPackedBitVector& parentRowMask);
-        //! Only called by split but is public so it's accessible to std::make_shared.
-        CLeafNodeStatistics(std::size_t id,
-                            const CLeafNodeStatistics& parent,
-                            const CLeafNodeStatistics& sibling,
-                            const TRegularization& regularization,
-                            const TSizeVec& featureBag,
-                            core::CPackedBitVector rowMask);
-
-        CLeafNodeStatistics(const CLeafNodeStatistics&) = delete;
-        CLeafNodeStatistics& operator=(const CLeafNodeStatistics&) = delete;
-
-        // Move construction/assignment not possible due to const reference member.
-
-        //! Apply the split defined by \p split.
-        //!
-        //! \return Shared pointers to the left and right child node statistics.
-        auto split(std::size_t leftChildId,
-                   std::size_t rightChildId,
-                   std::size_t numberThreads,
-                   const core::CDataFrame& frame,
-                   const CDataFrameCategoryEncoder& encoder,
-                   const TRegularization& regularization,
-                   const TImmutableRadixSetVec& candidateSplits,
-                   const TSizeVec& featureBag,
-                   const CBoostedTreeNode& split,
-                   bool leftChildHasFewerRows);
-
-        //! Order two leaves by decreasing gain in splitting them.
-        bool operator<(const CLeafNodeStatistics& rhs) const {
-            return m_BestSplit < rhs.m_BestSplit;
-        }
-
-        //! Get the gain in loss of the best split of this leaf.
-        double gain() const { return m_BestSplit.s_Gain; }
-
-        //! Get the total curvature of node.
-        double curvature() const { return m_BestSplit.s_Curvature; }
-
-        //! Get the best (feature, feature value) split.
-        TSizeDoublePr bestSplit() const {
-            return {m_BestSplit.s_Feature, m_BestSplit.s_SplitAt};
-        }
-
-        //! Check if the left child has fewer rows than the right child.
-        bool leftChildHasFewerRows() const {
-            return m_BestSplit.s_LeftChildHasFewerRows;
-        }
-
-        //! Check if we should assign the missing feature rows to the left child
-        //! of the split.
-        bool assignMissingToLeft() const {
-            return m_BestSplit.s_AssignMissingToLeft;
-        }
-
-        //! Get the node's identifier.
-        std::size_t id() const { return m_Id; }
-
-        //! Get the row mask for this leaf node.
-        core::CPackedBitVector& rowMask() { return m_RowMask; }
-
-        //! Get the memory used by this object.
-        std::size_t memoryUsage() const {
-            std::size_t mem{core::CMemory::dynamicSize(m_RowMask)};
-            mem += core::CMemory::dynamicSize(m_Derivatives);
-            mem += core::CMemory::dynamicSize(m_MissingDerivatives);
-            return mem;
-        }
-
-        //! Estimate the maximum leaf statistics' memory usage training on a data frame
-        //! with \p numberRows rows and \p numberCols columns using \p featureBagFraction
-        //! and \p numberSplitsPerFeature.
-        static std::size_t estimateMemoryUsage(std::size_t numberRows,
-                                               std::size_t numberCols,
-                                               std::size_t numberSplitsPerFeature) {
-            // We will typically get the close to the best compression for most of the
-            // leaves when the set of splits becomes large, corresponding to the worst
-            // case for memory usage. This is because the rows will be spread over many
-            // rows so the masks will mainly contain 0 bits in this case.
-            std::size_t rowMaskSize{numberRows / PACKED_BIT_VECTOR_MAXIMUM_ROWS_PER_BYTE};
-            std::size_t derivativesSize{(numberCols - 1) * numberSplitsPerFeature *
-                                        sizeof(SAggregateDerivatives)};
-            std::size_t missingDerivativesSize{(numberCols - 1) * sizeof(SAggregateDerivatives)};
-            return sizeof(CLeafNodeStatistics) + rowMaskSize + derivativesSize + missingDerivativesSize;
-        }
-
-    private:
-        using TDouble1Vec = core::CSmallVector<double, 1>;
-
-        //! \brief Statistics relating to a split of the node.
-        struct SSplitStatistics : private boost::less_than_comparable<SSplitStatistics> {
-            SSplitStatistics() = default;
-            SSplitStatistics(double gain,
-                             double curvature,
-                             std::size_t feature,
-                             double splitAt,
-                             bool leftChildHasFewerRows,
-                             bool assignMissingToLeft)
-                : s_Gain{gain}, s_Curvature{curvature}, s_Feature{feature}, s_SplitAt{splitAt},
-                  s_LeftChildHasFewerRows{leftChildHasFewerRows}, s_AssignMissingToLeft{assignMissingToLeft} {
-            }
-
-            bool operator<(const SSplitStatistics& rhs) const {
-                return COrderings::lexicographical_compare(
-                    s_Gain, s_Curvature, s_Feature, rhs.s_Gain, rhs.s_Curvature, rhs.s_Feature);
-            }
-
-            std::string print() const {
-                std::ostringstream result;
-                result << "split feature '" << s_Feature << "' @ " << s_SplitAt
-                       << ", gain = " << s_Gain;
-                return result.str();
-            }
-
-            double s_Gain = -INF;
-            double s_Curvature = 0.0;
-            std::size_t s_Feature = -1;
-            double s_SplitAt = INF;
-            bool s_LeftChildHasFewerRows = true;
-            bool s_AssignMissingToLeft = true;
-        };
-
-        //! \brief Aggregate derivatives.
-        struct SAggregateDerivatives {
-            void add(std::size_t count, const TDouble1Vec& gradient, const TDouble1Vec& curvature) {
-                s_Count += count;
-                s_Gradient += gradient[0];
-                s_Curvature += curvature[0];
-            }
-
-            void merge(const SAggregateDerivatives& other) {
-                this->add(other.s_Count, {other.s_Gradient}, {other.s_Curvature});
-            }
-
-            std::string print() const {
-                std::ostringstream result;
-                result << "count = " << s_Count << ", gradient = " << s_Gradient
-                       << ", curvature = " << s_Curvature;
-                return result.str();
-            }
-
-            std::size_t s_Count = 0;
-            double s_Gradient = 0.0;
-            double s_Curvature = 0.0;
-        };
-
-        using TAggregateDerivativesVec = std::vector<SAggregateDerivatives>;
-        using TAggregateDerivativesVecVec = std::vector<TAggregateDerivativesVec>;
-
-        //! \brief A collection of aggregate derivatives for candidate feature splits.
-        struct SSplitAggregateDerivatives {
-            SSplitAggregateDerivatives(const TImmutableRadixSetVec& candidateSplits)
-                : s_Derivatives(candidateSplits.size()),
-                  s_MissingDerivatives(candidateSplits.size()) {
-                for (std::size_t i = 0; i < candidateSplits.size(); ++i) {
-                    s_Derivatives[i].resize(candidateSplits[i].size() + 1);
-                }
-            }
-
-            void merge(const SSplitAggregateDerivatives& other) {
-                for (std::size_t i = 0; i < s_Derivatives.size(); ++i) {
-                    for (std::size_t j = 0; j < s_Derivatives[i].size(); ++j) {
-                        s_Derivatives[i][j].merge(other.s_Derivatives[i][j]);
-                    }
-                    s_MissingDerivatives[i].merge(other.s_MissingDerivatives[i]);
-                }
-            }
-
-            auto move() {
-                return std::make_pair(std::move(s_Derivatives),
-                                      std::move(s_MissingDerivatives));
-            }
-
-            TAggregateDerivativesVecVec s_Derivatives;
-            TAggregateDerivativesVec s_MissingDerivatives;
-        };
-
-    private:
-        void computeAggregateLossDerivatives(std::size_t numberThreads,
-                                             const core::CDataFrame& frame,
-                                             const CDataFrameCategoryEncoder& encoder);
-        void computeRowMaskAndAggregateLossDerivatives(std::size_t numberThreads,
-                                                       const core::CDataFrame& frame,
-                                                       const CDataFrameCategoryEncoder& encoder,
-                                                       bool isLeftChild,
-                                                       const CBoostedTreeNode& split,
-                                                       const core::CPackedBitVector& parentRowMask);
-
-        void addRowDerivatives(const CEncodedDataFrameRowRef& row,
-                               SSplitAggregateDerivatives& splitAggregateDerivatives) const;
-
-        SSplitStatistics computeBestSplitStatistics(const TRegularization& regularization,
-                                                    const TSizeVec& featureBag) const;
-
-    private:
-        std::size_t m_Id;
-        std::size_t m_Depth;
-        std::size_t m_NumberInputColumns;
-        std::size_t m_NumberLossParameters;
-        const TImmutableRadixSetVec& m_CandidateSplits;
-        core::CPackedBitVector m_RowMask;
-        TAggregateDerivativesVecVec m_Derivatives;
-        TAggregateDerivativesVec m_MissingDerivatives;
-        SSplitStatistics m_BestSplit;
-    };
-
-private:
-    // The maximum number of rows encoded by a single byte in the packed bit
-    // vector assuming best compression.
-    static const std::size_t PACKED_BIT_VECTOR_MAXIMUM_ROWS_PER_BYTE;
-    static const double INF;
 
 private:
     CBoostedTreeImpl();
@@ -577,7 +325,7 @@ private:
     TPackedBitVectorVec m_MissingFeatureRowMasks;
     TPackedBitVectorVec m_TrainingRowMasks;
     TPackedBitVectorVec m_TestingRowMasks;
-    double m_BestForestTestLoss = INF;
+    double m_BestForestTestLoss = boosted_tree_detail::INF;
     TOptionalDoubleVecVec m_FoldRoundTestLosses;
     CBoostedTreeHyperparameters m_BestHyperparameters;
     TNodeVecVec m_BestForest;
@@ -593,37 +341,6 @@ private:
 private:
     friend class CBoostedTreeFactory;
 };
-
-namespace boosted_tree_detail {
-inline std::size_t lossHessianStoredSize(std::size_t numberLossParameters) {
-    return numberLossParameters * (numberLossParameters + 1) / 2;
-}
-
-inline std::size_t numberLossParametersForHessianStoredSize(std::size_t lossHessianStoredSize) {
-    return static_cast<std::size_t>(
-        (std::sqrt(8.0 * static_cast<double>(lossHessianStoredSize) + 1.0) - 1.0) / 2.0 + 0.5);
-}
-
-inline std::size_t predictionColumn(std::size_t numberInputColumns) {
-    return numberInputColumns;
-}
-
-inline std::size_t lossGradientColumn(std::size_t numberInputColumns,
-                                      std::size_t numberLossParameters) {
-    return predictionColumn(numberInputColumns) + numberLossParameters;
-}
-
-inline std::size_t lossCurvatureColumn(std::size_t numberInputColumns,
-                                       std::size_t numberLossParameters) {
-    return lossGradientColumn(numberInputColumns, numberLossParameters) + numberLossParameters;
-}
-
-inline std::size_t exampleWeightColumn(std::size_t numberInputColumns,
-                                       std::size_t numberLossParameters) {
-    return lossCurvatureColumn(numberInputColumns, numberLossParameters) +
-           lossHessianStoredSize(numberLossParameters);
-}
-}
 }
 }
